% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gruendl_all_regex_terms}
\alias{gruendl_all_regex_terms}
\title{All regex terms by Gründl.}
\format{
A character vector
}
\usage{
gruendl_all_regex_terms
}
\description{
A character vector containing all the terms (excluding glob-style wildcards)
ever considered for inclusion in Gründl's German populism dictionary. You
should never have to load this vector directly.
}
\keyword{datasets}
