% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gruendl_terms_advocative_media_only}
\alias{gruendl_terms_advocative_media_only}
\title{Current terms in Gründl's dictionary for advocative populism in news media.}
\format{A character vector}
\usage{
gruendl_terms_advocative_media_only
}
\description{
A character vector containing all terms currently selected for inclusion in
Gründl's German populism dictionary for news media that indicate advocative
(for the people) populist messages. These terms do not include some of
the terms that only precisely hint at populist ideas if used by politicians
or parties.
}
\keyword{datasets}
