% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pauwels_2017_terms}
\alias{pauwels_2017_terms}
\title{Pauwels' (2017) dictionary terms.}
\format{A character vector}
\usage{
pauwels_2017_terms

pauwels_2017_terms
}
\description{
A character vector containing the terms suggested by Pauwels (2017).

A character vector containing the terms suggested by Pauwels (2017).
}
\references{
Pauwels, T. (2017). Measuring populism: A review of current approaches. In R. Heinisch, C. Holtz-Bacha, & O. Mazzoleni (Eds.), \emph{Political populism: A handbook} (pp. 123–136). Baden-Baden: Nomos.

Pauwels, T. (2017). Measuring populism: A review of current approaches. In R. Heinisch, C. Holtz-Bacha, & O. Mazzoleni (Eds.), \emph{Political populism: A handbook} (pp. 123–136). Baden-Baden: Nomos.
}
\keyword{datasets}
