% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quanteda-defaults.R
\name{pop_tokens_to_dfm}
\alias{pop_tokens_to_dfm}
\title{Transform tokens to a quanteda dfm object}
\usage{
pop_tokens_to_dfm(tokens, stopwords = NULL)
}
\arguments{
\item{tokens}{A quanteda tokens object, usually generated by the
get_pop_tokens provided with this package.}

\item{stopwords}{Stopwords are words or tokens which are removed from the
dfm. Defaults to NULL.}
}
\value{
A quanteda dfm object.
}
\description{
This function transforms a tokens object to a quanteda dfm
object. It makes some settings which ensure that dfm creation always uses
the same parameters. However, most of the settings are already made during
token creation in the function get_pop_tokens. This function is mainly for
transforming such a tokens object to a dfm object. It is basically a wrapper
for quanteda::dfm().
}
