% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rooduijn_2011_german_terms}
\alias{rooduijn_2011_german_terms}
\title{Rooduijn and Pauwels' (2011) dictionary terms.}
\format{
A character vector
}
\usage{
rooduijn_2011_german_terms
}
\description{
A character vector containing the terms suggested by Rooduijn and Pauwels
(2017).
}
\references{
Rooduijn, M., & Pauwels, T. (2011). Measuring populism: Comparing two methods of content analysis. \emph{West European Politics}, 34(6), 1272–1283. \href{https://doi.org/10.1080/01402382.2011.616665}{https://doi.org/10.1080/01402382.2011.616665}
}
\keyword{datasets}
