% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gruendl_terms}
\alias{gruendl_terms}
\title{Current terms in Gründl's populism dictionary.}
\format{A character vector}
\usage{
gruendl_terms
}
\description{
A character vector containing all terms currently selected for inclusion in
Gründl's German populism dictionary.
}
\keyword{datasets}
