% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gruendl_terms_conflictive}
\alias{gruendl_terms_conflictive}
\title{Current terms in Gründl's dictionary for conflictive populism.}
\format{
A character vector
}
\usage{
gruendl_terms_conflictive
}
\description{
A character vector containing all terms currently selected for inclusion in
Gründl's German populism dictionary that indicate conflictive (against the
elites) populist messages.
}
\keyword{datasets}
